refactor(server): split match.ts — extract 4 lifecycle handlers (S27.8.20)
Fifth slice on `apps/server/src/routes/match.ts`. Extract the 4
match-lifecycle handlers into a dedicated module
`routes/match-lifecycle-handlers.ts` (179 lines).

Endpoints covered:
- `POST /create` — handleCreateMatch (creates pending match + token)
- `POST /join` — handleJoinMatch (connects user + token)
- `POST /accept` — handleAcceptMatch (delegates to
  acceptAndMaybeStartMatch service)
- `POST /:id/cancel` — handleCancelMatch (delegates to cancelMatch
  service)

All thematically cohesive (mutate match lifecycle state) and isolated
from the move/state/details handlers that remain in `match.ts`.

Re-export from `match.ts` to preserve the public API consumed by
integration tests.

Leaf dependencies only (`prisma`, `jwt`, `MATCH_SECRET`,
`acceptAndMaybeStartMatch`, `cancelMatch`, `serverLog`,
`sendError/sendSuccess`). No cycle.

`match.ts` shrinks 809 -> 705 lines (-104 net lines, cumulative -1061
vs origin 1766 since S27.8.16 = -60%).

- 1787/1787 server tests pass (no regression)
- typecheck clean

Remaining in `match.ts`: handlePracticeMatch, handleSubmitMove +
4 inline setup/kickoff routes.
